+++ lib/src/application/phone/phone.dart
@@ -1,2 +1,1 @@
-export 'phone_input_session.dart';
 export 'phone_output_mapper.dart';